refactor: rename int_grid_image to impl_definitions
